a successful checkpoint of qwen3 fp32 implementation

+++ transformer/include/utils.h
@@ -28,10 +28,10 @@
 #define INT_ERROR_MAX 1e-5
 
 template <typename T>
-void read_to_array(const char* path, T* array, int size);
+void read_to_array(const char* path, T* array, size_t size);
 
 template <typename T>
-void write_array_to_file(const char* path, T* array, int size);
+void write_array_to_file(const char* path, T* array, size_t size);
 
 template <typename T>
 bool check_two_equal(T* array, T* array2, int size);
